+++ halma/src/Game/Halma/Configuration.hs
@@ -1,56 +1,41 @@
-{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveFunctor #-}
 {-# LANGUAGE DeriveTraversable #-}
 {-# LANGUAGE ExistentialQuantification #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
-{-# LANGUAGE KindSignatures #-}
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE StandaloneDeriving #-}
 
 module Game.Halma.Configuration
   ( HalmaPlayers (..)
-  , twoPlayers, threePlayers, fourPlayers, fivePlayers, sixPlayers
-  , getPlayers
-  , Configuration (..)
-  , defaultConfiguration
-  , SomeConfiguration (..)
+  , Configuration, configurationGrid, configurationPlayers
+  , configuration
+  , twoPlayersOnSmallGrid, threePlayersOnSmallGrid
+  , playersOnLargeGrid
+  , setSmallGrid, setLargeGrid
+  , addPlayerToConfig
+  , newGame
   ) where
 
 import Game.Halma.Board
+import Game.TurnCounter
 
 import Data.Aeson ((.=), (.:))
 import Data.Foldable (toList)
 import qualified Data.Aeson as A
-import qualified Data.Aeson.Types as A
 
-data HalmaPlayers :: HalmaGridSize -> * -> * where
-  TwoPlayers   :: a -> a ->                     HalmaPlayers size a
-  ThreePlayers :: a -> a -> a ->                HalmaPlayers size a
-  FourPlayers  :: a -> a -> a -> a ->           HalmaPlayers 'L   a
-  FivePlayers  :: a -> a -> a -> a -> a ->      HalmaPlayers 'L   a
-  SixPlayers   :: a -> a -> a -> a -> a -> a -> HalmaPlayers 'L   a
+data HalmaPlayers a
+  = TwoPlayers a a
+  | ThreePlayers a a a
+  | FourPlayers a a a a
+  | FivePlayers a a a a a
+  | SixPlayers a a a a a a
+  deriving (Eq, Show, Functor, Foldable, Traversable)
 
-deriving instance Eq a => Eq (HalmaPlayers size a)
-deriving instance Show a => Show (HalmaPlayers size a)
-deriving instance Functor (HalmaPlayers size)
-deriving instance Foldable (HalmaPlayers size)
-deriving instance Traversable (HalmaPlayers size)
-
-instance A.ToJSON a => A.ToJSON (HalmaPlayers size a) where
+instance A.ToJSON a => A.ToJSON (HalmaPlayers a) where
   toJSON = A.toJSON . toList
 
-instance A.FromJSON a => A.FromJSON (HalmaPlayers 'S a) where
-  parseJSON val = do
-    parsedPlayers <- A.parseJSON val
-    case parsedPlayers of
-      [a,b] -> pure (TwoPlayers a b)
-      [a,b,c] -> pure (ThreePlayers a b c)
-      _ ->
-        fail $ "unexpected count of players for a small board: " ++ show (length parsedPlayers)
-
-instance A.FromJSON a => A.FromJSON (HalmaPlayers 'L a) where
+instance A.FromJSON a => A.FromJSON (HalmaPlayers a) where
   parseJSON val = do
     parsedPlayers <- A.parseJSON val
     case parsedPlayers of
@@ -60,17 +45,9 @@ instance A.FromJSON a => A.FromJSON (HalmaPlayers 'L a) where
       [a,b,c,d,e] -> pure (FivePlayers a b c d e)
       [a,b,c,d,e,f] -> pure (SixPlayers a b c d e f)
       _ ->
-        fail $ "unexpected count of players for a large board: " ++ show (length parsedPlayers)
+        fail $ "unexpected count of players for a Halma board: " ++ show (length parsedPlayers)
 
-twoPlayers, threePlayers :: HalmaPlayers size ()
-twoPlayers = TwoPlayers () ()
-threePlayers = ThreePlayers () () ()
-fourPlayers, fivePlayers, sixPlayers :: HalmaPlayers 'L ()
-fourPlayers = FourPlayers () () () ()
-fivePlayers = FivePlayers () () () () ()
-sixPlayers = SixPlayers () () () () () ()
-
-getPlayers :: HalmaPlayers size a -> [(Team, a)]
+getPlayers :: HalmaPlayers a -> [(Team, a)]
 getPlayers halmaPlayers =
   case halmaPlayers of
     TwoPlayers a b ->
@@ -84,46 +61,77 @@ getPlayers halmaPlayers =
     SixPlayers a b c d e f ->
       [(North, a), (Northeast, b), (Southeast, c), (South, d), (Southwest, e), (Northwest, f)]
 
-data Configuration size a
+data Configuration a
   = Configuration
-  { configurationGrid :: HalmaGrid size
-  , configurationPlayers :: HalmaPlayers size a
+  { configurationGrid :: HalmaGrid
+  , configurationPlayers :: HalmaPlayers a
   } deriving (Eq, Show, Functor, Foldable, Traversable)
 
-instance A.ToJSON a => A.ToJSON (Configuration size a) where
+configuration :: HalmaGrid -> HalmaPlayers a -> Maybe (Configuration a)
+configuration grid players =
+  if grid /= SmallGrid || length players <= 3 then
+    Just
+      Configuration
+        { configurationGrid = grid
+        , configurationPlayers = players
+        }
+  else
+    Nothing
+
+instance A.ToJSON a => A.ToJSON (Configuration a) where
   toJSON config =
     A.object
       [ "grid" .= configurationGrid config
       , "players" .= configurationPlayers config
       ]
 
-parseConfigurationFromJSON
-  :: (A.FromJSON a, A.FromJSON (HalmaGrid size), A.FromJSON (HalmaPlayers size a))
-  => A.Value
-  -> A.Parser (Configuration size a)
-parseConfigurationFromJSON =
-  A.withObject "Configuration size" $ \o -> do
-    grid <- o .: "grid"
-    players <- o .: "players"
-    pure $
-      Configuration
-        { configurationGrid = grid
-        , configurationPlayers = players
-        }
+instance A.FromJSON a => A.FromJSON (Configuration a) where
+  parseJSON =
+    A.withObject "Configuration" $ \o -> do
+      grid <- o .: "grid"
+      players <- o .: "players"
+      case configuration grid players of
+        Nothing -> fail "too many players for small grid!"
+        Just config -> pure config
 
-instance A.FromJSON a => A.FromJSON (Configuration 'S a) where
-  parseJSON = parseConfigurationFromJSON
+twoPlayersOnSmallGrid :: a -> a -> Configuration a
+twoPlayersOnSmallGrid a b = Configuration SmallGrid (TwoPlayers a b)
 
-instance A.FromJSON a => A.FromJSON (Configuration 'L a) where
-  parseJSON = parseConfigurationFromJSON
+threePlayersOnSmallGrid :: a -> a -> a -> Configuration a
+threePlayersOnSmallGrid a b c = Configuration SmallGrid (ThreePlayers a b c)
 
-defaultConfiguration :: Configuration 'S ()
-defaultConfiguration = Configuration SmallGrid twoPlayers
+playersOnLargeGrid :: HalmaPlayers a -> Configuration a
+playersOnLargeGrid players = Configuration LargeGrid players
 
-data SomeConfiguration a = forall size. SomeConfiguration (Configuration size a)
+setSmallGrid :: Configuration a -> Maybe (Configuration a)
+setSmallGrid config =
+  if length (configurationPlayers config) <= 3 then
+    Just (config { configurationGrid = SmallGrid })
+  else
+    Nothing
 
-instance Show a => Show (SomeConfiguration a) where
-  show (SomeConfiguration a) = "SomeConfiguration (" ++ show a ++ ")"
+setLargeGrid :: Configuration a -> Configuration a
+setLargeGrid config =
+  config { configurationGrid = LargeGrid }
 
-instance Show a => Eq (SomeConfiguration a) where
-  SomeConfiguration a == SomeConfiguration b = show a == show b
+addPlayerToConfig :: a -> Configuration a -> Configuration a
+addPlayerToConfig new config =
+  case configurationPlayers config of
+    TwoPlayers a b ->
+      Configuration (configurationGrid config) (ThreePlayers a b new)
+    ThreePlayers a b c ->
+      Configuration LargeGrid (FourPlayers a b c new)
+    FourPlayers a b c d ->
+      Configuration LargeGrid (FivePlayers a b c d new)
+    FivePlayers a b c d e ->
+      Configuration LargeGrid (SixPlayers a b c d e new)
+    SixPlayers {} -> config
+
+newGame :: Configuration a -> (HalmaBoard, TurnCounter (Team, a))
+newGame config =
+  ( initialBoard (configurationGrid config) isActive
+  , newTurnCounter parties
+  )
+  where
+    parties = getPlayers (configurationPlayers config)
+    isActive color = color `elem` (fst <$> parties)