+++ halma/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Introduce new type `Move` representing moves on the board
 - changed call signature of `movePiece`
 - nicer default colours
+- removed `size :: GridSize` type index from `HalmaGrid` and `HalmaBoard`
 
 ## 0.2.0.1 (2015-07-22)